src/test/java/org/kowalsky/bankingapi/clinet/AlphaClientTest.java
package org.kowalsky.bankingapi.clinet;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.io.IOException;
import java.net.http.HttpClient;
import java.net.http.HttpRequest;
import java.net.http.HttpResponse;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.is;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class AlphaClientTest {

    @InjectMocks
    AlphaClient client;

    @Mock
    HttpClient httpClient;

    @Test
    public void test_getCurrencies_returnsResponse() throws IOException, InterruptedException {
        String testBody = "Test Body";
        HttpResponse<String> httpResponse = mock(HttpResponse.class);
        when(httpResponse.body()).thenReturn(testBody);
        when(httpClient.send(any(HttpRequest.class), any(HttpResponse.BodyHandler.class)))
                .thenReturn(httpResponse);

        assertThat(client.getCurrencies(), is(equalTo(testBody)));
    }

    @Test
    public void test_getCurrencies_throwsException() throws IOException, InterruptedException {
        when(httpClient.send(any(HttpRequest.class), any(HttpResponse.BodyHandler.class)))
                .thenThrow(IOException.class);

        assertThrows(RuntimeException.class, () -> client.getCurrencies());
    }
}